Registration: have user verify their emails
Registration process

1. User fills out form
2.* User receives email to verify their email address
3. Admin approves and activates users

*Admin can bypass this step and directly approve people if they want

+++ classes/Registration.php
@@ -154,18 +154,19 @@ class Registration
                 $user_id = $this->db_connection->lastInsertId();
 
                 if ($query_new_user_insert) {
-                    // Instead of sending a verification email, we let admin approve the users
-                    // if ($this->sendVerificationEmail($user_id, $user_email, $user_activation_hash)) {
+                    // We send a verification email for the user to verify their email address
+                    // The admin will activate the account once this is done
+                    if ($this->sendVerificationEmail($user_id, $user_email, $user_activation_hash)) {
                         $this->messages[] = MESSAGE_VERIFICATION_MAIL_SENT;
                         $this->registration_successful = true;
-                    /* } else {
+                    } else {
                         // delete this users account immediately, as we could not send a verification email
                         $query_delete_user = $this->db_connection->prepare('DELETE FROM users WHERE user_id=:user_id');
                         $query_delete_user->bindValue(':user_id', $user_id, PDO::PARAM_INT);
                         $query_delete_user->execute();
 
                         $this->errors[] = MESSAGE_VERIFICATION_MAIL_ERROR;
-                    }*/
+                    }
                 } else {
                     $this->errors[] = MESSAGE_REGISTRATION_FAILED;
                 }
@@ -218,6 +219,28 @@ class Registration
             return false;
         } else {
             return true;
+        }
+    }
+    
+    /**
+    * checks the id/verification code combination and removes the activation hash
+    */
+    public function verifyNewUser($user_id, $user_activation_hash)
+    {
+        // if database connection opened
+        if ($this->databaseConnection()) {
+            // try to update user with specified information
+            $query_update_user = $this->db_connection->prepare('UPDATE users SET user_activation_hash = NULL WHERE user_id = :user_id AND user_activation_hash = :user_activation_hash');
+            $query_update_user->bindValue(':user_id', intval(trim($user_id)), PDO::PARAM_INT);
+            $query_update_user->bindValue(':user_activation_hash', $user_activation_hash, PDO::PARAM_STR);
+            $query_update_user->execute();
+
+            if ($query_update_user->rowCount() > 0) {
+                $this->verification_successful = true;
+                $this->messages[] = MESSAGE_REGISTRATION_VERIFICATION_SUCCESSFUL;
+            } else {
+                $this->errors[] = MESSAGE_REGISTRATION_VERIFICATION_NOT_SUCCESSFUL;
+            }
         }
     }
 }
